api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-firestore
openai
pydantic
modal